requirements.txt
pytest~=8.4.2
playwright~=1.55.0
pytest-playwright
allure-pytest
pytest-html
Faker~=40.1.0